+++ flutter_ui_bloc/lib/flutter_ui_bloc.dart
@@ -10,3 +10,5 @@ export 'src/form/fields/base/DurationFieldBlocBuilder.dart';
 export 'src/form/fields/base/InputFileFieldBlocBuilder.dart';
 export 'src/form/fields/base/SliderFieldBlocBuilder.dart';
 export 'src/form/fields/base/upload_file_field_bloc_builder.dart';
+export 'src/form/fields/single/tab_bar_field_bloc_builders.dart';
+export 'src/form/typedefs.dart';