+++ flutter_ui_bloc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## [0.4.3]
+You can use one of these widgets to build your ui based on the `TabBar` widget and the `SelectFieldBloc`:
+- TabBarControllerFieldBlocProvider
+- TabBarFieldBlocBuilder
+- TabBarViewFieldBlocBuilder
+
+## [0.4.2]
+- First implementation of [UploadFileFieldBlocBuilder]
+
 ## [0.4.2]
 - First implementation of [UploadFileFieldBlocBuilder]